[WIP] tuning wifi again
Okay, going to tighten the thresholds up *just slightly* because it was pretty unusable, while still prefering 5GHz.
The < 800 estimated throughput scorer seems to be the best way to indicate preference for 5GHz networks.
Making it less prone to roam once it has a good network score regardless again, since it was disconnecting some people pretty frequently.
Not as iron-fisted grasp on current network as before, but 1 hour is a lot more reasonable than either 3 minutes (for testing) or 8 hours (wtf default).