sdm845-common: wifi: Update WCNSS_qcom_cfg.ini for WPA3 support
GCMP support was broken in older LA.UM.x.3 tags of kernel wifi driver
for sdm845 on Android 11 platform,but seems to be working in the newer
version that bananafunction imported.

Enable gcmp_enabled config in INI so that we stop choking on protected
management frames, without having to disable processing them
(and force people to downgrade their AP security settings).