sdm845-common: overlay: Setup ActivityManager parameters for A15
Trying to make sure we can still multitask on Android.

My usage often involves switching between a few 3-5 apps periodically, to answer an email or a message or check the calendar or quickly use calculator, and then go back to browser or game or "what I was doing" before.

Having the "what I was doing before" app get killed or compacted or otherwise forced to reload within 2-3 minutes of switching away is the obnoxious behavior I'm trying to mitigate.

* Attempt to increase max_cached_processes, which won't work because of anti-"abuse" commits in our fwb, but it's the principle of the thing
  (AOSP tries to prioritize all their stupid background crap like AI and Assistant over actual apps people WANT to use, and this is one of the only tuning knobs)
* Increase number of "Least Recently Used" (LRU) list apps up to 16, up from default of 8
* Enabled LRU app list re-prioritization
* Increased number of LRU apps to re-prioritize up to 5, up from default of 3
* Allow am to do proactive kills of LRU list apps if free swap falls below 5% (since it's not getting utilized as much otherwise)
* Increase background settle time up to 5 minutes, up from default of 1 minute